Bugfix for Hover which is visible above other application windows. #2534
This patch is fixing the problem, where the HoverManagers listeners are
deregistered but the hover is still visibile on top. This happens when
the hover was sticky eclipse was minimized and put back on foreground
after it. In that case the listeners are already deregistered and there
is no chance to react at all. So for this case the visibility has to be
set explicitly when the close is going to stop.
Version bump(s) for 4.35 stream
Bugfix for Hover which is visible above other application windows. #2534